# Conflicts:
#	app/src/main/res/layout/fragment_login.xml

+++ app/src/main/java/com/example/mainactivity/ui/account/AccountFragment.kt
@@ -10,17 +10,22 @@ import android.graphics.drawable.BitmapDrawable
 import android.net.Uri
 import android.os.Bundle
 import android.provider.MediaStore
+import android.text.Editable
+import android.text.TextUtils
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
+import android.widget.EditText
 import android.widget.ImageView
 import android.widget.TextView
 import android.widget.Toast
 import androidx.core.net.toFile
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
+import com.example.mainactivity.MainActivity
+import com.example.mainactivity.R
 import com.example.mainactivity.WelcomeActivity
 import com.example.mainactivity.databinding.FragmentAccountBinding
 import com.google.firebase.auth.FirebaseAuth
@@ -35,38 +40,54 @@ import com.google.firebase.storage.StorageReference
 import com.google.firebase.storage.ktx.storage
 import java.io.ByteArrayOutputStream
 import java.io.File
+import kotlin.time.Duration.Companion.nanoseconds
 
 class AccountFragment : Fragment() {
     private lateinit var auth: FirebaseAuth
     private var _binding: FragmentAccountBinding? = null
     private var pp:String = ""
+
     // This property is only valid between onCreateView and
     // onDestroyView.
     private val binding get() = _binding!!
     private lateinit var deconnect : Button
+    private lateinit var confirmer : Button
+
     private val pickImage = 100
     private var imageUri: Uri? = null
-    @SuppressLint("WrongThread")
+    private lateinit var prenom: EditText
+    private lateinit var nom: EditText
+
+    @SuppressLint("WrongThread", "MissingInflatedId")
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
         auth = FirebaseAuth.getInstance()
+        val view =  inflater.inflate(R.layout.fragment_account, container, false)
+        prenom = view.findViewById(R.id.prenom)
+        nom = view.findViewById(R.id.nom)
 
 
         val docRef = Firebase.firestore.collection("User")
         docRef.whereEqualTo("Mail",auth.currentUser?.email.toString()).get().addOnSuccessListener {
             result->
             for (document in result){
+
                 val gg = document.data["PP"].toString()
                 gg?.let {
                     // Assign the value to the global pp variable
                     pp = it
 
+                    var n = document.data["LastName"].toString()
+                    var p = document.data["FirstName"].toString()
+                    binding.prenom.setText(p)
+                    binding.nom.setText(n)
+
                     Thread.sleep(1000)
-                    Toast.makeText(getActivity(),pp ,
-                        Toast.LENGTH_LONG).show();
+
+
                     val StoRef = FirebaseStorage.getInstance().reference.child(pp)
                     val localfile = File.createTempFile(pp.split("/")[1].split(".")[0],pp.split("/")[1].split(".")[1],)
                     StoRef.getFile(localfile).addOnSuccessListener {
@@ -90,7 +111,6 @@ class AccountFragment : Fragment() {
         _binding = FragmentAccountBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textAccount
         binding.imageViewpp.setOnClickListener{
             val gallery = Intent(Intent.ACTION_PICK, MediaStore.Images.Media.INTERNAL_CONTENT_URI)
             startActivityForResult(gallery, pickImage)
@@ -110,6 +130,34 @@ class AccountFragment : Fragment() {
             startActivity(intent)
             requireActivity().finish()
 
+        }
+
+        //TESTS
+        confirmer = binding.confirmer
+        confirmer.setOnClickListener{
+            val txtPrenom = binding.prenom.text
+            val txtNom = binding.nom.text
+            val docRef = Firebase.firestore.collection("User")
+
+            if (TextUtils.isEmpty(txtNom) || TextUtils.isEmpty(txtPrenom)) //Vérification que les champs ne sont pas vides
+                Toast.makeText(
+                    activity,
+                    "Champs vides !",
+                    Toast.LENGTH_SHORT
+                ).show()
+
+            else
+                docRef.whereEqualTo("Mail",auth.currentUser?.email.toString()).get().addOnSuccessListener { result ->
+                    for (document in result) {
+                        docRef.document(document.id).update("FirstName", txtPrenom.toString())
+                        docRef.document(document.id).update("LastName", txtNom.toString())
+
+                }
+                    Toast.makeText(getActivity(), txtPrenom,
+                        Toast.LENGTH_LONG).show();
+                    //val intent = Intent(activity, MainActivity::class.java)
+                    //startActivity(intent)
+            }
         }
 
         return root
@@ -134,7 +182,12 @@ class AccountFragment : Fragment() {
             docRef.whereEqualTo("Mail",auth.currentUser?.email.toString()).get().addOnSuccessListener { result ->
                 for (document in result) {
                     val desertRef = storageRef.child(document.data["PP"].toString())
-                    desertRef.delete()
+                    if (document.data["PP"].toString().contentEquals("images/tele.jpeg"))
+                        Toast.makeText(getActivity(), "Image de base non supprimée",
+                            Toast.LENGTH_LONG).show();
+                    else{
+                        desertRef.delete()
+                    }
                     docRef.document(document.id).update("PP", "images/${file.lastPathSegment}.jpeg")
                 }
             }
@@ -146,6 +199,7 @@ class AccountFragment : Fragment() {
             }.addOnSuccessListener { taskSnapshot ->
 
             }
+
 
         }
     }
